requirements.txt
# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine learning
scikit-learn>=1.3.0
umap-learn>=0.5.3
hdbscan>=0.8.33
xgboost>=3.1.2  # Gradient boosting (CPU/GPU)

# Statistical analysis
statsmodels>=0.14.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
kaleido>=0.2.1  # For plotly static image export

# Progress bars
tqdm>=4.65.0

# Model persistence
joblib>=1.3.0

# Optional: for better performance
numba>=0.57.0

# Optional: for advanced interpretability
shap>=0.42.0  # Feature importance explanations
